Update station.py
(If you're Chloe and you're reading this) I figured out what went wrong with the in-app syncing. I've got 2 copies of Silver Pancake on my profile page. One is the old one, which I can't edit, and one is the new one.
Guess who downloaded the old one? ✋
(If you're the instructor and reading this) Howdy? 🤠